+++ weightress/src/main/java/iak/wrc/domain/entity/Weight.kt
@@ -1,3 +1,3 @@
 package iak.wrc.domain.entity
 
-data class Weight(val weight: Float, val notes: String, val date: Int)
+data class Weight(val weight: Float, val notes: String, val date: Long)
